refactor: Move Compilable de amora.types para amora.protocols
