+++ amora/protocols.py
@@ -1,6 +1,8 @@
-from typing import Protocol, runtime_checkable
+from typing import Protocol, Union, runtime_checkable
 
-from amora.types import Compilable
+from sqlalchemy.sql import Select, Selectable
+
+Compilable = Union[Select, Selectable]
 
 
 @runtime_checkable
