--- amora/types.py
@@ -1,5 +1,0 @@
-from typing import Union
-
-from sqlalchemy.sql import Select, Selectable
-
-Compilable = Union[Select, Selectable]
