+++ examples/amora_project/models/heart_agg.py
@@ -1,7 +1,7 @@
 from sqlmodel import Field, func, select
 
 from amora.models import AmoraModel, MaterializationTypes, ModelConfig
-from amora.types import Compilable
+from amora.protocols import Compilable
 from examples.amora_project.models.heart_rate import HeartRate
